# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ruckus Conference Program Operations API
  version: 1.0.0
  summary: Conference program operations - CFP, review, speakers, schedule, comms, CRM and public program feeds.
  description: >-
    HTTP API for Ruckus, an open-source conference program-operations tool. Paths are
    generated from the live Hono route registrations in src/app.ts and the src route
    modules, so path and method coverage is exact. Request and response bodies are
    intentionally loose object schemas with descriptions: handlers accept and return
    JSON envelopes shaped as a data property, and the demo does not enforce strict
    field level typing. Identity is demo persona simulation, not authentication.
  license:
    name: MIT
servers:
  - url: https://ruckus.to
    description: Live demo deployment
  - url: http://localhost:8787
    description: Local development API
...
security: []
paths:
...
  /api/crm/campaigns:
    get:
      tags: [crm]
      operationId: get_api_crm_campaigns
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/crm/communicate:
    post:
      tags: [crm]
      operationId: post_api_crm_communicate
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/crm/contacts:
    get:
      tags: [crm]
      operationId: get_api_crm_contacts
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      parameters:
        - name: company
          in: query
          required: false
          schema:
            type: string
          description: "company filter parameter."
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: "q filter parameter."
        - name: stage
          in: query
          required: false
          schema:
            type: string
          description: "stage filter parameter."
        - name: tag
          in: query
          required: false
          schema:
            type: string
          description: "tag filter parameter."
        - name: tags
          in: query
          required: false
          schema:
            type: string
          description: "tags filter parameter."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [crm]
      operationId: post_api_crm_contacts
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/crm/contacts/merge:
    post:
      tags: [crm]
      operationId: post_api_crm_contacts_merge
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/crm/contacts/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: "id path parameter."
    get:
      tags: [crm]
      operationId: get_api_crm_contacts_id
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [crm]
      operationId: patch_api_crm_contacts_id
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [crm]
      operationId: delete_api_crm_contacts_id
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/crm/contacts/{id}/add-to-event":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: "id path parameter."
    post:
      tags: [crm]
      operationId: post_api_crm_contacts_id_add_to_event
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/crm/contacts/{id}/notes":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: "id path parameter."
    post:
      tags: [crm]
      operationId: post_api_crm_contacts_id_notes
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/crm/contacts/{id}/stage":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: "id path parameter."
    post:
      tags: [crm]
      operationId: post_api_crm_contacts_id_stage
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/crm/dashboard:
    get:
      tags: [crm]
      operationId: get_api_crm_dashboard
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/crm/import:
    post:
      tags: [crm]
      operationId: post_api_crm_import
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/crm/import/validate:
    post:
      tags: [crm]
      operationId: post_api_crm_import_validate
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/crm/pipeline:
    get:
      tags: [crm]
      operationId: get_api_crm_pipeline
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  "/api/crm/segments/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: "id path parameter."
    delete:
      tags: [crm]
      operationId: delete_api_crm_segments_id
      summary: "Registered in crmRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
      description: >-
        Loosely typed JSON request body. Handlers read the fields they need and ignore
        the rest; see docs/API.md for the fields each operation consumes.
    DataEnvelope:
      type: object
      additionalProperties: true
      description: >-
        Standard success envelope. Most JSON endpoints return a data property holding an
        object or array; some also return meta describing counts or provider outcomes.
      properties:
        data:
          description: Operation payload, either an object or an array.
        meta:
          type: object
          additionalProperties: true
          description: Optional counts, filters or provider status details.
    ErrorEnvelope:
      type: object
      additionalProperties: true
      description: Standard failure envelope returned by the shared error helper.
      properties:
        error:
          type: object
          additionalProperties: true
          properties:
            code:
              type: string
              description: Short machine readable code such as NOT_FOUND or VALIDATION_ERROR.
            message:
              type: string
              description: Human readable explanation.